"Remove organization and user subscription services and related resources: delete controllers, services, repositories, DTOs, entities, and database migration scripts; replace with abstract billing subscription services for streamlined plan management."

+++ src/main/java/com/brainbyte/easy_maintenance/billing/application/dto/ChangePlanResponse.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 public record ChangePlanResponse(
     PlanChangeType type,
     Long invoiceId,
-    BigDecimal amountCharged,
+    Integer amountCharged,
     String effectiveAt
 ) {
     public enum PlanChangeType { UPGRADE, DOWNGRADE }
